+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Language server
 - Snippets
     - Open-ended timespan, `tsoe`/`timespan-open-ended`
     - Current time, `time`